feat: rename ThemeProvider to OrbitProvider
BREAKING CHANGE: ThemeProvider is now called OrbitProvider.
Everything else didn't change
