+++ .github/theming.md
@@ -2,13 +2,13 @@
 
 Our entire CSS styling for our components is based on [orbit-design-tokens](https://github.com/kiwicom/orbit-design-tokens), which contains variables with colors, sizings, spacings, etc. It also contains the functionality to create custom themes that can be used inside `orbit-components`.
 
-All you need to do is pass colors into the `getTokens` function and then pass this object into `<ThemeProvider />`. The component re-exports `ThemeProvider` from `styled-components` and will do all the magic for you thanks to React's context API.
+All you need to do is pass colors into the `getTokens` function and then pass this object into `<OrbitProvider />`. The component re-exports `OrbitProvider` from `styled-components` and will do all the magic for you thanks to React's context API.
 
 **Example:**
 
 ```jsx
 import getTokens from "@kiwicom/orbit-components/lib/getTokens";
-import ThemeProvider from "@kiwicom/orbit-components/lib/ThemeProvider";
+import OrbitProvider from "@kiwicom/orbit-components/lib/OrbitProvider";
 
 const customTokens = getTokens({
   palette: {
@@ -25,7 +25,7 @@ const customTokens = getTokens({
 });
 
 const App = () => (
-  <ThemeProvider
+  <OrbitProvider
     theme={{
       yourCustomTheme: { black: "#000" },
       orbit: customTokens,
@@ -35,13 +35,13 @@ const App = () => (
     }}
   >
     <Button type="secondary" size="large" />
-  </ThemeProvider>
+  </OrbitProvider>
 );
 ```
 
 ## Theme
 
-Table below contains all types of the props available in the ThemeProvider component.
+Table below contains all types of the props available in the OrbitProvider component.
 
 | Name                | Type      | Default | Description                                                                                            |
 | :------------------ | :-------- | :------ | :----------------------------------------------------------------------------------------------------- |
