+++ docs/src/documentation/01-getting-started/02-for-developers.mdx
@@ -54,8 +54,8 @@ Don't forget to also install [styled-components](https://github.com/styled-compo
    ```
 
 If you want to use a custom theme or dictionary inside your project,
-it's necessary to wrap your app in `<ThemeProvider>`.
-See [more info about the theme provider](https://github.com/kiwicom/orbit/tree/master/packages/orbit-components/src/ThemeProvider/README.md).
+it's necessary to wrap your app in `<OrbitProvider>`.
+See [more info about the Orbit provider](https://github.com/kiwicom/orbit/tree/master/packages/orbit-components/src/OrbitProvider/README.md).
 
 For a live preview, check out our [Storybook](https://kiwicom.github.io/orbit/) or [orbit.kiwi](https://orbit.kiwi).
