+++ docs/src/documentation/02-foundation/10-typography/02-circular-pro.mdx
@@ -39,7 +39,7 @@ It's actually not much harder, just a few more steps .
 
 ```jsx
 import getTokens from "@kiwicom/orbit-components/lib/getTokens"
-import ThemeProvider from "@kiwicom/orbit-components/lib/ThemeProvider"
+import OrbitProvider from "@kiwicom/orbit-components/lib/OrbitProvider"
 const customTokens = getTokens({
   base: {
     fontFamily: `"Circular Pro", -apple-system, ".SFNSText-Regular", "San Francisco",
@@ -47,9 +47,9 @@ const customTokens = getTokens({
   }
 })
 
-<ThemeProvider theme={{orbit: customTokens}}>
+<OrbitProvider theme={{orbit: customTokens}}>
   <App />
-</ThemeProvider>
+</OrbitProvider>
 ```
 
 ### In Figma
