+++ docs/styled.d.ts
@@ -1,7 +1,7 @@
 import "styled-components";
-import { ThemeProvider } from "@kiwicom/orbit-components";
+import { OrbitProvider } from "@kiwicom/orbit-components";
 
-type ThemeShape = React.ComponentProps<typeof ThemeProvider>["theme"];
+type ThemeShape = React.ComponentProps<typeof OrbitProvider>["theme"];
 
 // https://styled-components.com/docs/api#typescript
