+++ docs/wrapWithProviders.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ThemeProvider } from "@kiwicom/orbit-components";
+import { OrbitProvider } from "@kiwicom/orbit-components";
 
 import { DevModeProvider } from "./src/hooks/useDevMode";
 import { TableOfContentsProvider } from "./src/services/table-of-contents";
@@ -8,13 +8,13 @@ import { KeyboardContextProvider } from "./src/services/KeyboardProvider";
 
 export const wrapWithProviders = ({ element }) => {
   return (
-    <ThemeProvider theme={theme}>
+    <OrbitProvider theme={theme}>
       <KeyboardContextProvider>
         <DevModeProvider>
           <TableOfContentsProvider>{element}</TableOfContentsProvider>
         </DevModeProvider>
       </KeyboardContextProvider>
-    </ThemeProvider>
+    </OrbitProvider>
   );
 };
