+++ packages/orbit-components/src/OrbitProvider/README.md
@@ -1,22 +1,22 @@
-# ThemeProvider
+# OrbitProvider
 
-orbit-components has theming support via our own `<ThemeProvider>` which adds you possibilities to add your own theme and dictionary.
+orbit-components has theming support via our own `<OrbitProvider>` which adds you possibilities to add your own theme and dictionary.
 
 ```jsx
-import ThemeProvider from "@kiwicom/orbit-components/lib/ThemeProvider";
+import OrbitProvider from "@kiwicom/orbit-components/lib/OrbitProvider";
 ```
 
-After adding import please wrap your application into `ThemeProvider` and you can provide your own [`theme`](https://github.com/kiwicom/orbit/blob/master/.github/theming.md) and [`dictionary`](https://github.com/kiwicom/orbit/blob/master/docs/src/documentation/05-development/01-guides/02-dictionary.mdx)
+After adding import please wrap your application into `OrbitProvider` and you can provide your own [`theme`](https://github.com/kiwicom/orbit/blob/master/.github/theming.md) and [`dictionary`](https://github.com/kiwicom/orbit/blob/master/docs/src/documentation/05-development/01-guides/02-dictionary.mdx)
 
 ```jsx
-<ThemeProvider theme={ownTheme} dictionary={ownDictionary}>
+<OrbitProvider theme={ownTheme} dictionary={ownDictionary}>
   <App />
-</ThemeProvider>
+</OrbitProvider>
 ```
 
 ## Props
 
-Table below contains all types of the props available in the ThemeProvider component.
+Table below contains all types of the props available in the OrbitProvider component.
 
 | Name         | Type         | Default | Description                                                                                                                         |
 | :----------- | :----------- | :------ | :---------------------------------------------------------------------------------------------------------------------------------- |
