+++ packages/orbit-components/src/OrbitProvider/ThemeProvider.stories.tsx
@@ -9,7 +9,7 @@ import languages from "../data/dictionary";
 import Badge from "../Badge";
 import Tooltip from "../Tooltip";
 
-import ThemeProvider from ".";
+import OrbitProvider from ".";
 
 const ButtonWithTranslation = () => {
   const translate = useTranslate();
@@ -17,15 +17,15 @@ const ButtonWithTranslation = () => {
 };
 
 export default {
-  title: "ThemeProvider",
+  title: "OrbitProvider",
 };
 
 export const DictionaryContext = () => {
   const dictionary = select("dictionary", Object.keys(languages), "en");
   return (
-    <ThemeProvider theme={{ orbit: getTokens() }} dictionary={languages[dictionary]}>
+    <OrbitProvider theme={{ orbit: getTokens() }} dictionary={languages[dictionary]}>
       <ButtonWithTranslation />
-    </ThemeProvider>
+    </OrbitProvider>
   );
 };
 
@@ -41,11 +41,11 @@ DictionaryContext.story = {
 export const WithoutTransitions = () => {
   const transitions = boolean("transitions", false);
   return (
-    <ThemeProvider theme={{ orbit: getTokens(), transitions }}>
+    <OrbitProvider theme={{ orbit: getTokens(), transitions }}>
       <Tooltip content="Lorem ipsum dolor sit amet">
         <Badge>Info</Badge>
       </Tooltip>
-    </ThemeProvider>
+    </OrbitProvider>
   );
 };
 
@@ -77,9 +77,9 @@ export const OwnTheme = () => {
   });
   const customTheme = object("customTheme", { black: "#000" });
   return (
-    <ThemeProvider theme={{ orbit: { ...getTokens(orbitTheme), ...customTheme } }}>
+    <OrbitProvider theme={{ orbit: { ...getTokens(orbitTheme), ...customTheme } }}>
       <Button onClick={action("onClick")}>Hello World!</Button>
-    </ThemeProvider>
+    </OrbitProvider>
   );
 };
