+++ packages/orbit-components/src/hooks/useMediaQuery/__tests__/index.test.tsx
@@ -3,7 +3,7 @@ import { render, act } from "@testing-library/react";
 import { renderToString } from "react-dom/server";
 import MatchMediaMock from "jest-matchmedia-mock";
 
-import ThemeProvider from "../../../ThemeProvider";
+import OrbitProvider from "../../../OrbitProvider";
 import defaultTheme from "../../../defaultTheme";
 import useMediaQuery from "..";
 
@@ -39,13 +39,13 @@ describe("useMediaQuery", () => {
 
     function App() {
       return (
-        <ThemeProvider theme={theme}>
+        <OrbitProvider theme={theme}>
           <MediaQuery
             onChange={query => {
               result = query;
             }}
           />
-        </ThemeProvider>
+        </OrbitProvider>
       );
     }
 
@@ -109,7 +109,7 @@ describe("useMediaQuery", () => {
     matchMedia.useMediaQuery(`(min-width: ${widthBreakpointMediumMobile}px)`);
 
     render(
-      <ThemeProvider
+      <OrbitProvider
         theme={{
           ...theme,
           orbit: {
@@ -123,7 +123,7 @@ describe("useMediaQuery", () => {
             result = query;
           }}
         />
-      </ThemeProvider>,
+      </OrbitProvider>,
     );
 
     expect(result).toMatchInlineSnapshot(`
@@ -176,18 +176,18 @@ describe("useMediaQuery", () => {
       const matchMedia = new MatchMediaMock();
 
       const { rerender } = render(
-        <ThemeProvider theme={theme}>
+        <OrbitProvider theme={theme}>
           <UpdateCounter />
-        </ThemeProvider>,
+        </OrbitProvider>,
       );
 
       expect(updateCount).toBe(2);
 
       rerender(
         // change theme object reference to trigger re-render
-        <ThemeProvider theme={{ ...theme }}>
+        <OrbitProvider theme={{ ...theme }}>
           <UpdateCounter />
-        </ThemeProvider>,
+        </OrbitProvider>,
       );
 
       // nothing changed, so query object shouldn't either
@@ -198,16 +198,16 @@ describe("useMediaQuery", () => {
       });
 
       rerender(
-        <ThemeProvider theme={{ ...theme }}>
+        <OrbitProvider theme={{ ...theme }}>
           <UpdateCounter />
-        </ThemeProvider>,
+        </OrbitProvider>,
       );
 
       // viewport changed, query object should have updated
       expect(updateCount).toBe(3);
 
       rerender(
-        <ThemeProvider
+        <OrbitProvider
           theme={{
             ...theme,
             orbit: {
@@ -217,7 +217,7 @@ describe("useMediaQuery", () => {
           }}
         >
           <UpdateCounter />
-        </ThemeProvider>,
+        </OrbitProvider>,
       );
 
       // breakpoint no longer matches the viewport, query object should have updated
