+++ packages/orbit-components/src/utils/rtl/RenderInRtl.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { defaultTokens } from "@kiwicom/orbit-design-tokens";
 
-import ThemeProvider from "../../ThemeProvider";
+import OrbitProvider from "../../OrbitProvider";
 
 interface Props {
   readonly children: React.ReactNode;
@@ -24,9 +24,9 @@ class RenderInRtl extends React.PureComponent<Props> {
 
   render() {
     return (
-      <ThemeProvider theme={{ orbit: defaultTokens, rtl: true }}>
+      <OrbitProvider theme={{ orbit: defaultTokens, rtl: true }}>
         {this.props.children}
-      </ThemeProvider>
+      </OrbitProvider>
     );
   }
 }
