+++ packages/orbit-themer/src/App.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import assocPath from "ramda/src/assocPath";
 import { hot } from "react-hot-loader";
 import styled, { createGlobalStyle } from "styled-components";
-import { getTokens, ThemeProvider } from "@kiwicom/orbit-components";
+import { getTokens, OrbitProvider } from "@kiwicom/orbit-components";
 
 import Components from "./Components";
 import Tabs from "./Tabs";
@@ -57,7 +57,7 @@ const App = () => {
         }}
       >
         <>
-          <ThemeProvider
+          <OrbitProvider
             theme={{
               orbit: getTokens({
                 base: {
@@ -68,8 +68,8 @@ const App = () => {
             }}
           >
             <Tabs />
-          </ThemeProvider>
-          <ThemeProvider
+          </OrbitProvider>
+          <OrbitProvider
             theme={{
               orbit: getTokens({
                 palette: colors,
@@ -81,7 +81,7 @@ const App = () => {
             }}
           >
             <Components />
-          </ThemeProvider>
+          </OrbitProvider>
         </>
       </ColorContext.Provider>
     </StyledApp>
